IntelliJ: Use single quote instead of double quote for sync query string in Windows.
In Windows (Powershell), double quote will allow the string (in this case, `^((?!manual).)*$`) to be evaluated. That string has some special meaning in Windows and is unexpectedly evaluated to something we don't want. We instead, want that string to be passed as-is.

In this case, we should use single quote which is universally used for that purpose. For safety purpose, this is currently only enabled for Windows.

PiperOrigin-RevId: 412263075