Set up a Scala SDK that includes more than just the Scala Library and fix header_ bug from rules_jvm_external.
The Scala 2 SDK was quite simple and included essentially just a library
and compiler. The Scala 3 SDK has many more dependencies.

This is a quick and hacky way of setting this up. Long term, this should be
modified such that the SDK jars are pulled from the Scala toolchain.

This also fixes the bug caused by rules_jvm_external prepending header_
to the compile jars. The header_ prefix breaks some things in the IntelliJ
Scala plugin, so long term, we should work with the rules_jvm_external
folks to fix that.